pyproject.toml
[tool.poetry]
name = "Centrilearn"
version = "0.1.0"
description = "Centrilearn Web Application"
authors = ["Emerald Henry"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
ipython = "^8.22.2"
python-dotenv = "^1.0.1"
django = "^4.2.2"
pillow = "^10.3.0"
asgiref = "^3.8.1"
autopep8 = "^2.1.0"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
pycodestyle = "^2.11.1"
pytz = "^2024.1"
sqlparse = "^0.5.0"
toml = "^0.10.2"
requests = "^2.31.0"
langchain = "^0.1.16"
llama-index = "^0.10.33"
pytest = "^8.2.0"
openai = "^1.23.6"
selenium = "^4.20.0"
ollama = "^0.1.9"
crewai = "^0.19.0"




[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"